ai-model/enterprise_computer_use/requirements.txt
streamlit>=1.38.0
anthropic[bedrock,vertex]>=0.37.1
openai==1.59.9
jsonschema==4.22.0
boto3>=1.28.57
google-auth<3,>=2
gymnasium==1.0.0
grpcio==1.70.0
grpcio-tools==1.70.0
PyAutoGUI==0.9.54
pyairtable==3.0.1
any-parser==0.0.20
google-api-python-client==2.160.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1

annotated-types==0.7.0
anyio==4.8.0
click==8.1.8
fastapi==0.115.11
h11==0.14.0
idna==3.10
pydantic==2.10.6
pydantic_core==2.27.2
sniffio==1.3.1
starlette==0.46.0
typing_extensions==4.12.2
uvicorn==0.34.0
